+++ R/S3-methods.R
@@ -35,7 +35,7 @@ print.sjmisc.frq <- function(x, ...) {
 print.sjmisc.descr <- function(x, ...) {
   cat("## Basic descriptive statistics\n\n")
   # round values
-  x[, c(4:6, 8, 12:13)] <- round(x[, c(4:6, 8, 12:13)], 2)
+  x[, c(4:6, 8, 12:14)] <- round(x[, c(4:6, 8, 12:14)], 2)
   # print frq-table
   print.data.frame(x, ..., row.names = FALSE)
 }