+++ NEWS.md
@@ -1,3 +1,10 @@
+# sjmisc 2.3.0.9000
+
+## Changes to functions
+
+* `descr()` gets a `max.length`-argument to shorten variable labels in the output to a specific number of chars.
+* `descr()` now also reports the percentage of missing values.
+
 # sjmisc 2.3.0
 
 ## General